home/dev-env/utilities/python/pyproject.toml
[project]
name = "python-utils"
version = "0.1.0"
description = "Diverse python utilities for everyday tasks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "ruff>=0.8.3",
]
